# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Talyn Gold Trading API",
        "description": "API for gold trading platform",
        "contact": {
            "name": "Talyn Support",
            "email": "info@talyn.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get all orders for the authenticated user",
                "operationId": "955bb77a45ae173a485fadb8befed93f",
                "responses": {
                    "200": {
                        "description": "List of orders",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "type": {
                                                "type": "string",
                                                "example": "buy"
                                            },
                                            "price": {
                                                "type": "integer",
                                                "example": 10000000
                                            },
                                            "amount": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 2
                                            },
                                            "remaining_amount": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 1.5
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "partial"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2023-01-01T12:00:00Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create a new order",
                "operationId": "2123ba58dadb22bd6f2c8081af562230",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "price",
                                    "amount"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "buy",
                                            "sell"
                                        ],
                                        "example": "buy"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": 10000000
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "type": {
                                            "type": "string",
                                            "example": "buy"
                                        },
                                        "price": {
                                            "type": "integer",
                                            "example": 10000000
                                        },
                                        "amount": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 2.5
                                        },
                                        "remaining_amount": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 2.5
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "open"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-01-01T12:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (e.g., insufficient balance)"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{id}": {
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Cancel an order",
                "operationId": "bab875a1ed962fb33565ab9b48c0d494",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order cancelled successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "cancelled"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Order cancelled successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (e.g., order already filled)"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized access to order"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}